requirements.txt
gymnasium==1.0.0
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.2
pandas==2.2.3
Pillow==11.1.0
tqdm==4.67.1
